Get collection metadata
List of changes introduced in this commit:
- Added OpenSea API client (using gotcha)
- Read collection metadata from OpenSea
- Added more utilities & tests
- Minor improvements & type updates

+++ pkg/database/types.go
@@ -35,7 +35,7 @@ const (
 )
 
 type Links struct {
-	Timestamp int    `firestore:"timestamp,omitempty"`
+	Timestamp int64  `firestore:"timestamp,omitempty"`
 	Twitter   string `firestore:"twitter,omitempty"`
 	Discord   string `firestore:"discord,omitempty"`
 	External  string `firestore:"external,omitempty"`
